Vincent's ReVengeance reVival
The Revenmity deck has finally been introduced, optimized and released at full capacity. Showing perhaps some of the most insane Ritual plays there are. With the capability of constantly reviving bodies for the ceremonial summon, as well as tapping on their restless strenght, you'll find this deck does not stay dead for long.